Scheduled and Forced Authority Set Change Handling

+++ src/test/java/com/limechain/sync/fullsync/InMemoryDB.java
@@ -49,7 +49,7 @@ class InMemoryDB implements KVRepository<String, Object> {
 
     @Override
     public DeleteByPrefixResult deleteByPrefix(String prefix, Long limit) {
-        var keys = this.findKeysByPrefix(prefix, limit.intValue());
+        var keys = this.findKeysByPrefix(prefix, Math.toIntExact(limit));
 
         int deleted = 0;
         for (var key : keys) {
@@ -57,7 +57,7 @@ class InMemoryDB implements KVRepository<String, Object> {
             deleted++;
         }
 
-        boolean allDeleted = this.findKeysByPrefix(prefix, limit.intValue()).isEmpty();
+        boolean allDeleted = this.findKeysByPrefix(prefix, Math.toIntExact(limit)).isEmpty();
 
         return new DeleteByPrefixResult(deleted, allDeleted);
     }
